[BUGFIX] Fix issue on UniqueFieldHandler
With TYPO3 13 the implementation of the FormPersistenceManager changed.
When trying to save into an extension and having `form_to_database`
active, this save errors.

This patch fixes this issue and a test will make sure, saving to
extensions will work in the future.

FOr further information take a look at
https://review.typo3.org/c/Packages/TYPO3.CMS/+/85669

Fixes #118